+++ src/main/java/racing_game/controller/RacingGameController.java
@@ -1,6 +1,10 @@
-package step3;
+package racing_game.controller;
 
-public class Main {
+import racing_game.model.Game;
+import racing_game.model.RacingGame;
+import racing_game.view.RacingGameInput;
+
+public class RacingGameController {
   public static void main(String[] args) {
     Game racingGame = new RacingGame(RacingGameInput.getNameOfCarsByScanner(),
         RacingGameInput.getTimeByScanner());
