+++ src/main/java/racing_game/model/Car.java
@@ -1,4 +1,4 @@
-package step3;
+package racing_game.model;
 
 public class Car implements Comparable<Car> {
