+++ src/main/java/racing_game/model/Game.java
@@ -1,4 +1,4 @@
-package step3;
+package racing_game.model;
 
 public interface Game {
   void play();
