+++ src/main/java/racing_game/model/RacingCarMovingStrategy.java
@@ -1,4 +1,4 @@
-package step3;
+package racing_game.model;
 
 public class RacingCarMovingStrategy implements CarMovingStartegy {
