+++ src/main/java/racing_game/model/RacingGameUtils.java
@@ -1,4 +1,4 @@
-package step3;
+package racing_game.model;
 
 import java.util.Random;
