+++ src/test/java/racing_game/model/RacingGameUtilsTest.java
@@ -1,8 +1,9 @@
-package step3;
+package racing_game.model;
 
 import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing_game.model.RacingGameUtils;
 
 class RacingGameUtilsTest {
